refactor(mcp): split surface policy and resource modules
